requirements.txt
fastapi
uvicorn
networkx
groq
python-dotenv
pydantic
